refactor(derive): compact macro module layout
