+++ derive/src/account/methods.rs
@@ -1,0 +1,152 @@
+use {
+    super::{dynamic, fixed::PodFieldInfo},
+    crate::helpers::zc_assign_from_value,
+    quote::{format_ident, quote},
+};
+
+pub(super) fn emit_set_inner_impl(
+    name: &syn::Ident,
+    vis: &syn::Visibility,
+    field_infos: &[PodFieldInfo<'_>],
+    has_dynamic: bool,
+    disc_len: usize,
+    zc_name: &syn::Ident,
+    zc_path: &proc_macro2::TokenStream,
+    gen_set_inner: bool,
+) -> proc_macro2::TokenStream {
+    if !gen_set_inner {
+        return quote! {};
+    }
+
+    if has_dynamic {
+        let inner_name = format_ident!("{}Inner", name);
+        let inner_fields: Vec<proc_macro2::TokenStream> = field_infos
+            .iter()
+            .map(|fi| {
+                let field_name = fi.field.ident.as_ref().expect("field must be named");
+                match &fi.pod_dyn {
+                    None => {
+                        let field_ty = &fi.field.ty;
+                        quote! { pub #field_name: #field_ty }
+                    }
+                    Some(dyn_field) => dynamic::emit_inner_field(field_name, dyn_field),
+                }
+            })
+            .collect();
+        let max_checks: Vec<proc_macro2::TokenStream> = field_infos
+            .iter()
+            .filter_map(|fi| {
+                let field_name = fi.field.ident.as_ref().expect("field must be named");
+                fi.pod_dyn
+                    .as_ref()
+                    .map(|dyn_field| dynamic::emit_max_check(field_name, dyn_field))
+            })
+            .collect();
+        let space_terms: Vec<proc_macro2::TokenStream> = field_infos
+            .iter()
+            .filter_map(|fi| {
+                let field_name = fi.field.ident.as_ref().expect("field must be named");
+                fi.pod_dyn
+                    .as_ref()
+                    .map(|dyn_field| dynamic::emit_space_term(field_name, dyn_field))
+            })
+            .collect();
+        let zc_header_stmts: Vec<proc_macro2::TokenStream> = field_infos
+            .iter()
+            .filter(|fi| fi.pod_dyn.is_none())
+            .map(|fi| {
+                zc_assign_from_value(
+                    fi.field.ident.as_ref().expect("field must be named"),
+                    &fi.field.ty,
+                )
+            })
+            .collect();
+        let var_write_stmts: Vec<proc_macro2::TokenStream> = field_infos
+            .iter()
+            .filter_map(|fi| {
+                let field_name = fi.field.ident.as_ref().expect("field must be named");
+                fi.pod_dyn
+                    .as_ref()
+                    .map(|dyn_field| dynamic::emit_write_stmt(field_name, dyn_field))
+            })
+            .collect();
+        let init_field_names: Vec<&syn::Ident> = field_infos
+            .iter()
+            .map(|fi| fi.field.ident.as_ref().expect("field must be named"))
+            .collect();
+
+        quote! {
+            #vis struct #inner_name<'a> {
+                #(#inner_fields,)*
+            }
+
+            impl #name {
+                #[inline(always)]
+                pub fn set_inner(
+                    &mut self,
+                    inner: #inner_name<'_>,
+                    payer: &AccountView,
+                    rent_lpb: u64,
+                    rent_threshold: u64,
+                ) -> Result<(), ProgramError> {
+                    #(let #init_field_names = inner.#init_field_names;)*
+                    #(#max_checks)*
+
+                    let __space = Self::MIN_SPACE #(#space_terms)*;
+                    let __view = unsafe { &mut *(self as *mut Self as *mut AccountView) };
+
+                    if __space != __view.data_len() {
+                        quasar_lang::accounts::account::realloc_account_raw(
+                            __view,
+                            __space,
+                            payer,
+                            rent_lpb,
+                            rent_threshold,
+                        )?;
+                    }
+
+                    let __ptr = __view.data_mut_ptr();
+                    let __zc = unsafe { &mut *(__ptr.add(#disc_len) as *mut #zc_name) };
+                    #(#zc_header_stmts)*
+                    let __dyn_start = #disc_len + core::mem::size_of::<#zc_name>();
+                    let __len = __view.data_len();
+                    let __data = unsafe {
+                        core::slice::from_raw_parts_mut(__ptr.add(__dyn_start), __len - __dyn_start)
+                    };
+                    let mut __offset = 0usize;
+                    #(#var_write_stmts)*
+                    let _ = __offset;
+                    Ok(())
+                }
+            }
+        }
+    } else {
+        let inner_name = format_ident!("{}Inner", name);
+        let field_names: Vec<_> = field_infos.iter().map(|fi| &fi.field.ident).collect();
+        let field_types: Vec<_> = field_infos.iter().map(|fi| &fi.field.ty).collect();
+        let set_inner_stmts: Vec<proc_macro2::TokenStream> = field_infos
+            .iter()
+            .map(|fi| {
+                zc_assign_from_value(
+                    fi.field.ident.as_ref().expect("field must be named"),
+                    &fi.field.ty,
+                )
+            })
+            .collect();
+
+        quote! {
+            #vis struct #inner_name {
+                #(pub #field_names: #field_types,)*
+            }
+
+            impl #name {
+                #[inline(always)]
+                pub fn set_inner(&mut self, inner: #inner_name) {
+                    let __zc = unsafe { &mut *(self.__view.data_mut_ptr().add(#disc_len) as *mut #zc_path) };
+                    #(let #field_names = inner.#field_names;)*
+                    #(#set_inner_stmts)*
+                }
+            }
+        }
+    }
+}
