+++ derive/src/account/traits.rs
@@ -1,0 +1,107 @@
+use {super::fixed::PodFieldInfo, crate::helpers::map_to_pod_type, quote::quote};
+
+pub(super) fn emit_discriminator_impl(
+    name: &syn::Ident,
+    disc_bytes: &[syn::LitInt],
+    bump_offset_impl: &proc_macro2::TokenStream,
+) -> proc_macro2::TokenStream {
+    quote! {
+        impl Discriminator for #name {
+            const DISCRIMINATOR: &'static [u8] = &[#(#disc_bytes),*];
+            #bump_offset_impl
+        }
+    }
+}
+
+pub(super) fn emit_owner_impl(name: &syn::Ident) -> proc_macro2::TokenStream {
+    quote! {
+        impl Owner for #name {
+            const OWNER: Address = crate::ID;
+        }
+    }
+}
+
+pub(super) fn emit_space_impl(
+    name: &syn::Ident,
+    field_infos: &[PodFieldInfo<'_>],
+    has_dynamic: bool,
+    disc_len: usize,
+    zc_path: &proc_macro2::TokenStream,
+    prefix_total: usize,
+) -> proc_macro2::TokenStream {
+    if has_dynamic {
+        quote! {
+            impl Space for #name {
+                const SPACE: usize = #disc_len + core::mem::size_of::<#zc_path>() + #prefix_total;
+            }
+        }
+    } else {
+        let field_pod_types: Vec<proc_macro2::TokenStream> = field_infos
+            .iter()
+            .map(|fi| map_to_pod_type(&fi.field.ty))
+            .collect();
+        quote! {
+            impl Space for #name {
+                const SPACE: usize = #disc_len #(+ core::mem::size_of::<#field_pod_types>())*;
+            }
+        }
+    }
+}
+
+pub(super) fn emit_account_check_impl(
+    name: &syn::Ident,
+    has_dynamic: bool,
+    disc_len: usize,
+    disc_indices: &[usize],
+    disc_bytes: &[syn::LitInt],
+    zc_path: &proc_macro2::TokenStream,
+    prefix_total: usize,
+    validation_stmts: &[proc_macro2::TokenStream],
+) -> proc_macro2::TokenStream {
+    if has_dynamic {
+        quote! {
+            impl AccountCheck for #name {
+                type Params = ();
+
+                #[inline(always)]
+                fn check(view: &AccountView) -> Result<(), ProgramError> {
+                    let __data = unsafe { view.borrow_unchecked() };
+                    let __data_len = __data.len();
+                    let __min = #disc_len + core::mem::size_of::<#zc_path>() + #prefix_total;
+                    if __data_len < __min {
+                        return Err(ProgramError::AccountDataTooSmall);
+                    }
+                    #(
+                        if unsafe { *__data.get_unchecked(#disc_indices) } != #disc_bytes {
+                            return Err(ProgramError::InvalidAccountData);
+                        }
+                    )*
+                    let mut __offset = #disc_len + core::mem::size_of::<#zc_path>();
+                    #(#validation_stmts)*
+                    let _ = __offset;
+                    Ok(())
+                }
+            }
+        }
+    } else {
+        quote! {
+            impl AccountCheck for #name {
+                type Params = ();
+
+                #[inline(always)]
+                fn check(view: &AccountView) -> Result<(), ProgramError> {
+                    let __data = unsafe { view.borrow_unchecked() };
+                    if __data.len() < #disc_len + core::mem::size_of::<#zc_path>() {
+                        return Err(ProgramError::AccountDataTooSmall);
+                    }
+                    #(
+                        if unsafe { *__data.get_unchecked(#disc_indices) } != #disc_bytes {
+                            return Err(ProgramError::InvalidAccountData);
+                        }
+                    )*
+                    Ok(())
+                }
+            }
+        }
+    }
+}
