+++ derive/src/accounts/syntax/mod.rs
@@ -1,0 +1,14 @@
+mod attrs;
+mod instruction_args;
+mod pda;
+
+pub(crate) use {
+    attrs::{parse_field_attrs, AccountDirective},
+    instruction_args::{
+        generate_instruction_arg_extraction, parse_struct_instruction_args, InstructionArg,
+    },
+    pda::{
+        classify_seed, lower_bump, render_seed_expr, seeds_to_emit_nodes, SeedEmitNode,
+        SeedRenderContext,
+    },
+};
